Conflicts:
	src/server/game/Entities/AreaTrigger/AreaTrigger.cpp
	src/server/game/Entities/AreaTrigger/AreaTrigger.h
	src/server/game/Entities/Creature/Creature.cpp
	src/server/game/Entities/DynamicObject/DynamicObject.cpp
	src/server/game/Entities/DynamicObject/DynamicObject.h
	src/server/game/Entities/GameObject/GameObject.cpp
	src/server/game/Entities/GameObject/GameObject.h
	src/server/game/Entities/Item/Container/Bag.cpp
	src/server/game/Entities/Item/Item.cpp
	src/server/game/Entities/Object/Object.cpp
	src/server/game/Entities/Object/Object.h
	src/server/game/Entities/Object/ObjectDefines.h
	src/server/game/Entities/Player/Player.cpp
	src/server/game/Entities/Player/Player.h
	src/server/game/Entities/Unit/Unit.cpp
	src/server/game/Entities/Unit/Unit.h
	src/server/game/Guilds/Guild.cpp
	src/server/game/Handlers/AuctionHouseHandler.cpp
	src/server/game/Handlers/ItemHandler.cpp
	src/server/game/Handlers/MiscHandler.cpp
	src/server/game/Handlers/PetitionsHandler.cpp
	src/server/game/Handlers/SpellHandler.cpp
	src/server/game/Handlers/TradeHandler.cpp
	src/server/game/Spells/Spell.cpp
	src/server/game/Spells/SpellEffects.cpp
	src/server/newscripts/EasternKingdoms/BastionOfTwilight/boss_halfus_wyrmbreaker.cpp
	src/server/newscripts/EasternKingdoms/Karazhan/bosses_opera.cpp
	src/server/newscripts/EasternKingdoms/MagistersTerrace/boss_felblood_kaelthas.cpp
	src/server/newscripts/EasternKingdoms/Uldaman/instance_uldaman.cpp
	src/server/newscripts/Pandaria/TempleJadeSerpent/instance_temple_of_jade_serpent.cpp
	src/server/scripts/Commands/cs_misc.cpp
	src/server/scripts/Kalimdor/BlackfathomDeeps/blackfathom_deeps.cpp
	src/server/scripts/Kalimdor/CavernsOfTime/BattleForMountHyjal/hyjalAI.cpp
	src/server/scripts/Kalimdor/CavernsOfTime/EndTime/boss_echo_of_sylvanas.cpp
	src/server/scripts/Kalimdor/CavernsOfTime/EndTime/boss_murozond.cpp
	src/server/scripts/Maelstrom/lost_isles.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheChampion/trial_of_the_champion.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheCrusader/instance_trial_of_the_crusader.cpp
	src/server/scripts/Northrend/FrozenHalls/HallsOfReflection/halls_of_reflection.cpp
	src/server/scripts/Northrend/FrozenHalls/HallsOfReflection/instance_halls_of_reflection.cpp
	src/server/scripts/Northrend/Gundrak/instance_gundrak.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_deathbringer_saurfang.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_the_lich_king.cpp
	src/server/scripts/Northrend/Naxxramas/boss_kelthuzad.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_razorscale.cpp
	src/server/scripts/Northrend/UtgardeKeep/UtgardePinnacle/instance_pinnacle.cpp
	src/server/scripts/Northrend/icecrown.cpp
	src/server/scripts/Outland/BlackTemple/boss_reliquary_of_souls.cpp
	src/server/scripts/Outland/TempestKeep/Mechanar/boss_gatewatcher_gyrokill.cpp
	src/server/scripts/Spells/spell_druid.cpp
	src/server/scripts/Spells/spell_monk.cpp
	src/server/scripts/Spells/spell_rogue.cpp

+++ src/server/shared/Logging/Log.h
@@ -71,7 +71,7 @@ class Log
         void outSpamm( const char * str, ... )               ATTR_PRINTF(2, 3);
         void outDiff( const char * str, ... )                ATTR_PRINTF(2, 3);
         void outCommand(uint32 account, const char * str, ...) ATTR_PRINTF(3, 4);
-        void outCharDump(char const* str, uint32 account_id, uint32 guid, char const* name);
+        void outCharDump(char const* str, uint32 account_id, uint64 guid, char const* name);
 
         void EnableDBAppenders();
         static std::string GetTimestampStr();
